Don't build persist.img
Files in the persist partition are created in the factory and contain
IDs and calibration data that are unique to every single device.

Overwriting the persist partition bricks a device. So remove persist.img
from the build through the 'FP3_SKIP_PERSIST_IMG' flag.

Issue: FP3-A11#31
Issue: FP3-A11#202
Change-Id: I9089a8c3d974248bb02fd8d56f12f49d924c00a8